Update Encounter.js
Z should be used at the end of a timestamp only as a meaning of UTC time: See ISO 8601.

"If the time is in UTC, add a Z directly after the time without a space. Z is the zone designator for the zero UTC offset. If no time zone is specified (by a 'Z' or a time offset), the time zone in which the date is expressed is supposed to be the local one." 

https://en.wikipedia.org/wiki/ISO_8601.